+++ implants/Cargo.toml
@@ -73,6 +73,7 @@ which = "4.4.2"
 whoami = "1.3.0"
 windows-sys = "0.45.0"
 winreg = "0.51.0"
+windows-service = "0.6.0"
 
 [profile.release]
 strip = true  # Automatically strip symbols from the binary.
